Separate app.bsky methods out into app view
