+++ packages/pds/src/app-view/api/app/bsky/actor/getProfiles.ts
@@ -1,5 +1,5 @@
-import { Server } from '../../../../lexicon'
-import AppContext from '../../../../context'
+import { Server } from '../../../../../lexicon'
+import AppContext from '../../../../../context'
 
 export default function (server: Server, ctx: AppContext) {
   server.app.bsky.actor.getProfiles({
