+++ packages/pds/src/app-view/api/app/bsky/actor/searchTypeahead.ts
@@ -1,22 +1,19 @@
-import { sql } from 'kysely'
-import AppContext from '../../../../context'
-import Database from '../../../../db'
-import { Server } from '../../../../lexicon'
-import * as Method from '../../../../lexicon/types/app/bsky/actor/search'
-import { getDeclarationSimple } from '../util'
+import AppContext from '../../../../../context'
+import Database from '../../../../../db'
+import { Server } from '../../../../../lexicon'
+import * as Method from '../../../../../lexicon/types/app/bsky/actor/search'
+import { getDeclarationSimple } from '../../../../../api/app/bsky/util'
 import {
   cleanTerm,
   getUserSearchQueryPg,
   getUserSearchQuerySqlite,
-  SearchKeyset,
-} from '../../../../services/util/search'
+} from '../../../../../services/util/search'
 
 export default function (server: Server, ctx: AppContext) {
-  server.app.bsky.actor.search({
+  server.app.bsky.actor.searchTypeahead({
     auth: ctx.accessVerifier,
     handler: async ({ params }) => {
       let { term, limit } = params
-      const { before } = params
 
       term = cleanTerm(term || '')
       limit = Math.min(limit ?? 25, 100)
@@ -32,64 +29,53 @@ export default function (server: Server, ctx: AppContext) {
 
       const results =
         ctx.db.dialect === 'pg'
-          ? await getResultsPg(ctx.db, { term, limit, before })
-          : await getResultsSqlite(ctx.db, { term, limit, before })
+          ? await getResultsPg(ctx.db, { term, limit })
+          : await getResultsSqlite(ctx.db, { term, limit })
 
       const users = results.map((result) => ({
         did: result.did,
         declaration: getDeclarationSimple(result),
         handle: result.handle,
         displayName: result.displayName ?? undefined,
-        description: result.description ?? undefined,
         avatar: result.avatarCid
           ? ctx.imgUriBuilder.getCommonSignedUri('avatar', result.avatarCid)
           : undefined,
-        indexedAt: result.indexedAt ?? undefined,
       }))
-
-      const keyset = new SearchKeyset(sql``, sql``)
 
       return {
         encoding: 'application/json',
         body: {
           users,
-          cursor: keyset.packFromResult(results),
         },
       }
     },
   })
 }
 
-const getResultsPg: GetResultsFn = async (db, { term, limit, before }) => {
-  return await getUserSearchQueryPg(db, { term: term || '', limit, before })
+const getResultsPg: GetResultsFn = async (db, { term, limit }) => {
+  return await getUserSearchQueryPg(db, { term: term || '', limit })
     .leftJoin('profile', 'profile.creator', 'did_handle.did')
     .select([
-      'distance',
       'did_handle.did as did',
-      'did_handle.handle as handle',
-      'did_handle.actorType as actorType',
       'did_handle.declarationCid as declarationCid',
+      'did_handle.actorType as actorType',
+      'did_handle.handle as handle',
       'profile.displayName as displayName',
-      'profile.description as description',
       'profile.avatarCid as avatarCid',
-      'profile.indexedAt as indexedAt',
     ])
     .execute()
 }
 
-const getResultsSqlite: GetResultsFn = async (db, { term, limit, before }) => {
-  return await getUserSearchQuerySqlite(db, { term: term || '', limit, before })
+const getResultsSqlite: GetResultsFn = async (db, { term, limit }) => {
+  return await getUserSearchQuerySqlite(db, { term: term || '', limit })
     .leftJoin('profile', 'profile.creator', 'did_handle.did')
     .select([
-      sql<number>`0`.as('distance'),
       'did_handle.did as did',
-      'did_handle.handle as handle',
-      'did_handle.actorType as actorType',
       'did_handle.declarationCid as declarationCid',
+      'did_handle.actorType as actorType',
+      'did_handle.handle as handle',
       'profile.displayName as displayName',
-      'profile.description as description',
       'profile.avatarCid as avatarCid',
-      'profile.indexedAt as indexedAt',
     ])
     .execute()
 }
@@ -100,13 +86,10 @@ type GetResultsFn = (
 ) => Promise<
   {
     did: string
-    actorType: string
     declarationCid: string
+    actorType: string
     handle: string
     displayName: string | null
-    description: string | null
     avatarCid: string | null
-    distance: number
-    indexedAt: string | null
   }[]
 >
