+++ packages/pds/src/app-view/api/app/bsky/feed/getRepostedBy.ts
@@ -1,8 +1,8 @@
-import { Server } from '../../../../lexicon'
-import { paginate, TimeCidKeyset } from '../../../../db/pagination'
-import { getDeclarationSimple } from '../util'
-import AppContext from '../../../../context'
-import { notSoftDeletedClause } from '../../../../db/util'
+import { Server } from '../../../../../lexicon'
+import { paginate, TimeCidKeyset } from '../../../../../db/pagination'
+import { getDeclarationSimple } from '../../../../../api/app/bsky/util'
+import AppContext from '../../../../../context'
+import { notSoftDeletedClause } from '../../../../../db/util'
 
 export default function (server: Server, ctx: AppContext) {
   server.app.bsky.feed.getRepostedBy({
